chromium/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/html/imports/import-readd-nested-expected.txt

CONSOLE WARNING: HTML Imports is deprecated and has now been removed as of M80. See https://www.chromestatus.com/features/5144752345317376 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
This is a testharness.js-based test.
FAIL Styles in grandchild are re-applied after removed imports are re-added assert_equals: expected "rgb(255, 0, 0)" but got "rgb(0, 0, 0)"
Harness: the test ran to completion.
